refactor: make `graph` submodule of `fragmenter`
