+++ src/meta/src/stream/mod.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 mod fragmenter;
-mod graph;
 mod meta;
 mod scheduler;
 mod source_manager;
